TAD cafeteras
Thanks to the c++ programming language, we are able to create an example based on an exercise related to the Abstract Data Types or TAD, where we can appreciate the abstraction of the coffee makers in this case Oster and Haceb. 

Class Coffee Maker:

Constructor: Initializes the name of the coffee maker.
Methods cappuccino and tinto: Simulates making coffee and displays a message.
Method getName: Returns the name of the coffee maker.

Main:

Initialization of coffeeMakerCurrent: It initially points to coffeemaker1 (Oster).
do-while cycle: Allows the user to select the type of coffee, change the brewer, or exit the program.
Coffee brewer change: The user can switch between Oster and Haceb coffee brewers.